refactor(git): Make git service package more consistent
