+++ services/git-service-api/src/main/java/io/fabric8/launcher/service/git/github/api/GitHubEnvVarSysPropNames.java
@@ -1,4 +1,4 @@
-package io.fabric8.launcher.service.github.api;
+package io.fabric8.launcher.service.git.github.api;
 
 /**
  * Contains names of environment variables or system properties
