+++ services/git-service-api/src/main/java/io/fabric8/launcher/service/git/gitlab/api/GitLabEnvVarSysPropNames.java
@@ -1,4 +1,4 @@
-package io.fabric8.launcher.service.gitlab.api;
+package io.fabric8.launcher.service.git.gitlab.api;
 
 /**
  * Contains names of environment variables or system properties
