+++ services/git-service-impl/src/main/java/io/fabric8/launcher/service/git/bitbucket/BitbucketServiceFactory.java
@@ -1,4 +1,4 @@
-package io.fabric8.launcher.service.bitbucket;
+package io.fabric8.launcher.service.git.bitbucket;
 
 import java.util.Optional;
 
@@ -7,12 +7,12 @@ import javax.enterprise.context.ApplicationScoped;
 import io.fabric8.launcher.base.EnvironmentSupport;
 import io.fabric8.launcher.base.identity.Identity;
 import io.fabric8.launcher.base.identity.IdentityFactory;
-import io.fabric8.launcher.service.bitbucket.api.BitbucketEnvVarSysPropNames;
+import io.fabric8.launcher.service.git.bitbucket.api.BitbucketEnvVarSysPropNames;
 import io.fabric8.launcher.service.git.api.GitServiceFactory;
 import io.fabric8.launcher.service.git.spi.GitProvider;
 
-import static io.fabric8.launcher.service.bitbucket.api.BitbucketEnvVarSysPropNames.LAUNCHER_MISSIONCONTROL_BITBUCKET_APPLICATION_PASSWORD;
-import static io.fabric8.launcher.service.bitbucket.api.BitbucketEnvVarSysPropNames.LAUNCHER_MISSIONCONTROL_BITBUCKET_USERNAME;
+import static io.fabric8.launcher.service.git.bitbucket.api.BitbucketEnvVarSysPropNames.LAUNCHER_MISSIONCONTROL_BITBUCKET_APPLICATION_PASSWORD;
+import static io.fabric8.launcher.service.git.bitbucket.api.BitbucketEnvVarSysPropNames.LAUNCHER_MISSIONCONTROL_BITBUCKET_USERNAME;
 import static io.fabric8.launcher.service.git.spi.GitProvider.GitProviderType.BITBUCKET;
 
 @ApplicationScoped
