+++ services/git-service-impl/src/main/java/io/fabric8/launcher/service/git/github/KohsukeGitHubRepository.java
@@ -1,4 +1,4 @@
-package io.fabric8.launcher.service.github;
+package io.fabric8.launcher.service.git.github;
 
 import java.net.URI;
 import java.net.URISyntaxException;
