+++ services/git-service-impl/src/main/java/io/fabric8/launcher/service/git/github/KohsukeGitHubWebhook.java
@@ -1,4 +1,4 @@
-package io.fabric8.launcher.service.github;
+package io.fabric8.launcher.service.git.github;
 
 import java.util.List;
 import java.util.Objects;
