+++ services/git-service-impl/src/main/java/io/fabric8/launcher/service/git/gitlab/GitLabServiceFactory.java
@@ -1,4 +1,4 @@
-package io.fabric8.launcher.service.gitlab;
+package io.fabric8.launcher.service.git.gitlab;
 
 import java.util.Optional;
 
@@ -10,10 +10,10 @@ import io.fabric8.launcher.base.identity.IdentityFactory;
 import io.fabric8.launcher.base.identity.TokenIdentity;
 import io.fabric8.launcher.service.git.api.GitServiceFactory;
 import io.fabric8.launcher.service.git.spi.GitProvider;
-import io.fabric8.launcher.service.gitlab.api.GitLabEnvVarSysPropNames;
+import io.fabric8.launcher.service.git.gitlab.api.GitLabEnvVarSysPropNames;
 
 import static io.fabric8.launcher.service.git.spi.GitProvider.GitProviderType.GITLAB;
-import static io.fabric8.launcher.service.gitlab.api.GitLabEnvVarSysPropNames.LAUNCHER_MISSIONCONTROL_GITLAB_PRIVATE_TOKEN;
+import static io.fabric8.launcher.service.git.gitlab.api.GitLabEnvVarSysPropNames.LAUNCHER_MISSIONCONTROL_GITLAB_PRIVATE_TOKEN;
 
 /**
  * @author <a href="mailto:ggastald@redhat.com">George Gastaldi</a>
